Implement basic tic tac toe gameplay
Add a Player class with initialize and attr_reader for the
variables name and marker. Add a Grid class with initialize, []
and []= methods for the rows variable, check methods for rows,
columns and diagonals, winner? and end? methods and layout methods.
Add a Game class with initialize method  which calls game_intro,
initialize_players, initialize_grid, main_rountine and game_end
methods; also add print_grid, get_input and valid_input methods.
The code also contains some comments describing each class.

The Player class stores the player related data. The Grid class
store the grid states, checks for a winner or end game and displays
the grid in a convenient layout for the command line. The Game
class controls the game, shows a game intro, initializes players
and grid, take players moves each turn checking for valid inputs
and outputs the game result when we have a winner or draw.